+++ graphics/r_renderer.h
@@ -11,6 +11,10 @@
 namespace Verse::Graphics::Renderer::GL
 {
     void create(Config &c, SDL_Window* window, ui8 &pid);
+
+    ui32 create_texture(Tex* tex, int w, int h);
+    void render_texture(ui32 &tex_id, Rect &src, Rect &dst, ui16 frames);
+
     void clear();
     void present(SDL_Window* window, ui8 &pid);
     void destroy();