+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Conditional compiling for SDL or OpenGL
 - Basic OpenGL Renderer, VAO and VBO
 - Version in the title
+- Texture rendering
+- Animation spritesheet support
 #### Changed
 - Code cleanup
 - Moved imgui to deps